Add type annotation to logical operators
