+++ src/typeInference/annotateAST.ts
@@ -1,7 +1,6 @@
 import { SequenceMultiplicity, SequenceType, ValueType } from '../expressions/dataTypes/Value';
 import StaticContext from '../expressions/StaticContext';
 import astHelper, { IAST } from '../parsing/astHelper';
-import { annotateAndOperator } from './annotateAndOperator';
 import { annotateBinOp } from './annotateBinaryOperator';
 import { annotateCastableOperator, annotateCastOperator } from './annotateCastOperators';
 import {
@@ -10,6 +9,7 @@ import {
 	annotateValueCompare,
 } from './annotateCompareOperator';
 import { annotateFunctionCall } from './annotateFunctionCall';
+import { annotateLogicalOperator } from './annotateLogicalOperator';
 import { annotateUnaryMinus, annotateUnaryPlus } from './annotateUnaryOperator';
 
 /**
@@ -68,11 +68,12 @@ export function annotate(ast: IAST, staticContext: StaticContext): SequenceType 
 			return annotateBinOp(ast, left, right, astNodeName);
 		}
 
-		//And + Or operators
+		// And + Or operators
 		case 'andOp':
+		case 'orOp':
 			annotate(astHelper.getFirstChild(ast, 'firstOperand')[1] as IAST, staticContext);
 			annotate(astHelper.getFirstChild(ast, 'secondOperand')[1] as IAST, staticContext);
-			return annotateAndOperator(ast);
+			return annotateLogicalOperator(ast);
 
 		// Comparison operators
 		case 'equalOp':
